move avatar with popup to separate component
