+++ ui/src/components/FeedMew.vue
@@ -1,27 +1,9 @@
 <template>
   <q-item-section avatar>
-    <agent-avatar
-      :agent-pub-key="authorPubKey(mew.feedMew.header.author)"
-      size="50"
-      class="self-start cursor-pointer"
-      @click="onAgentClick(authorPubKey(mew.feedMew.header.author))"
-      @mouseenter="showProfile(index)"
-      @mouseleave="hideProfile(index)"
-    >
-      <q-menu
-        v-model="profileVisible[index]"
-        anchor="bottom middle"
-        self="top middle"
-        :offset="[-10, 5]"
-        no-focus
-      >
-        <ProfilePopup
-          :agent-pub-key="authorPubKey(mew.feedMew.header.author)"
-          @mouseenter="keepShowingProfile(index)"
-          @mouseleave="hideProfile(index)"
-        />
-      </q-menu>
-    </agent-avatar>
+    <avatar-with-popup
+      :mew="mew"
+      :index="index"
+    />
   </q-item-section>
 
   <q-item-section>
@@ -42,7 +24,7 @@
     <div>
       <q-btn
         size="sm"
-        :icon="isLickedByMe() ? 'favorite' : 'favorite_border'"
+        :icon="isLickedByMe ? 'favorite' : 'favorite_border'"
         flat
         @click="toggleLickMew"
       >
@@ -91,7 +73,7 @@
         <AddMew
           class="text-center"
           :mew-type="{ reply: mew.mewEntryHash }"
-          @publish-mew="closeTextBox"
+          @publish-mew="publishReply"
         />
       </q-card-section>
     </q-card>
@@ -99,24 +81,21 @@
 </template>
 
 <script setup lang="ts">
-import { HoloHashB64 } from "@holochain-open-dev/core-types";
 import { lickMew, unlickMew } from "../services/clutter-dna";
 import { computed, onMounted, ref } from "vue";
 import {
-  FeedMewWithContext as FeedMew,
+  FeedMewWithContext,
   CreateMewInput,
 } from "../types/types";
-import { authorPubKey } from "../utils/hash";
-import { useRouter } from "vue-router";
 import AddMew from "./AddMew.vue";
 import MewsItemContent from "./MewsItemContent.vue";
-import ProfilePopup from "./ProfilePopup.vue";
 import { PropType } from "vue";
 import { useProfileStore } from "../services/profile-store";
 import Timestamp from "./Timestamp.vue";
+import AvatarWithPopup from "./AvatarWithPopup.vue";
 
-let props = defineProps({
-  mew: { type: Object as PropType<FeedMew>, required: true },
+const props = defineProps({
+  mew: { type: Object as PropType<FeedMewWithContext>, required: true },
   index: { type: Number, required: true }
 });
 const store = useProfileStore();
@@ -125,14 +104,6 @@ const displayName = computed(() => agentProfile.value?.fields["Display name"]);
 const nickname = computed(() => agentProfile.value?.nickname);
 const myAgentPubKey = store.myAgentPubKey;
 
-const PROFILE_SHOW_HIDE_DELAY = 400; // in ms
-
-const router = useRouter();
-
-const mewsFeed = ref<FeedMew[]>([]);
-const profileVisible = ref<boolean[]>([]);
-const profileHideTimeouts = ref<number[]>([]);
-const profileShowTimeouts = ref<number[]>([]);
 const emit = defineEmits<{
   (e: 'publish-mew', mew: CreateMewInput): void;
   (e: 'refresh-feed'): void;
@@ -142,35 +113,11 @@ onMounted(async () => {
   agentProfile.value = await store.fetchAgentProfile(props.mew.feedMew.header.author);
 });
 
-let isReplying = ref(false);
-let isMewMewing = ref(false);
-
-const onAgentClick = (agentPubKey: HoloHashB64) => {
-  router.push(`/profiles/${agentPubKey}`);
-};
-
-const showProfile = (index: number) => {
-  profileVisible.value = new Array(mewsFeed.value.length).fill(false);
-  profileShowTimeouts.value[index] = window.setTimeout(
-    () => (profileVisible.value[index] = true),
-    PROFILE_SHOW_HIDE_DELAY
-  );
-};
-
-const hideProfile = (index: number) => {
-  window.clearTimeout(profileShowTimeouts.value[index]);
-  profileHideTimeouts.value[index] = window.setTimeout(
-    () => (profileVisible.value[index] = false),
-    PROFILE_SHOW_HIDE_DELAY
-  );
-};
-
-const keepShowingProfile = (index: number) => {
-  window.clearTimeout(profileHideTimeouts.value[index]);
-};
+const isReplying = ref(false);
+const isLickedByMe = computed(() => props.mew.licks.includes(myAgentPubKey));
 
 const toggleLickMew = async () => {
-  if (isLickedByMe()) {
+  if (isLickedByMe.value) {
     await unlickMew(props.mew.mewEntryHash);
   } else {
     await lickMew(props.mew.mewEntryHash);
@@ -178,24 +125,14 @@ const toggleLickMew = async () => {
   emit("refresh-feed");
 };
 
-const isLickedByMe = () => props.mew.licks.includes(myAgentPubKey);
+const replyToMew = () => isReplying.value = true;
 
-const replyToMew = () => {
-  if (isMewMewing.value) {
-    mewMew();
-  }
-  isReplying.value = !isReplying.value;
-};
 const mewMew = () => {
-  if (isReplying.value) {
-    replyToMew();
-  }
-  isMewMewing.value = !isMewMewing.value;
+  console.log('hello');
 };
 
-const closeTextBox = (newMew: CreateMewInput) => {
+const publishReply = (newMew: CreateMewInput) => {
   isReplying.value = false;
-  isMewMewing.value = false;
   emit("publish-mew", newMew);
 };
 </script>
